Add files via upload
This is the code for my RNG Game and my RPS Game. With these I wasn't able to add the code which allowed the user to choose between both games. I also struggled with adding the code to add a point counter and a loop to play the RPS Game 5 times. Would you be able to go through this with me next session so I can understand what I need to add and improve on with this. Thank you.